BUGS; FreqCount84 now prints "DICK" out the UART pin repeatedly.
Currently, it does this REALLY reliably. I don't see any misprinted
characters in thousands upon thousands.
I am doing this with a delay. This will become a drag on the system's
display because it lowers the duty cycle of the seven-segment display. Or
maybe it will make the last digit brighter or something. I'm pretty sure
the DICK thing goes berserk if you actually try to measure a frequency.