FreqCount now says "FUCK" at 19200 baud. Repeatedly! Progress! I chose to disable all interrupts I was using when I print the frequency measurement. This ruins any current frequency measurement that was being made. The quality of the UART transmission is now very good. It is based on the delay functions. I'm thinking I could get this to work at higher bauds, but they might require some fine tuning of the delay functions.
Alternatively, now that all of the other irritating interrupt sources are gone, I could enable the only one I need: the timer1 compare b timer which gives me the exact time that bits should be shifted out. That might be promising.

However, 19200 is probably good enough most of the time.
For a 13-byte message, (10 bits each. start bit, byte, stop bit) it would take 10 * 13 * 52 us = 6.7 ms.

If I got this working at 115200, it would take only 1.1 ms.

One of the many still outstanding issues is the fact that the last (right-most) 7-segment digit is brighter than the others when a ton of UART communication is being performed.

Under normal operation, you probably won't have a shitload of UART frequency measurements happening so fast, but it would be cool if the display still worked correctly even when the UART Tx is being abused.